Latest version. The build succesfully completed and the test passed.
For a rerun before merging this in the master branch, the argument -b install should be added to the git clone.

Note that the build.sh was a little stripped because it contained some problems. As a result it must be noted that all additional apt-get commands should be after the apt-get update. This is also required since the Ubuntu 14.04 server contains very few packages by default.