Initial version of starter and install supporting docker.
As there is no vps available to test this over it is for now remaining as initial.

An important thing to check remains the checking if ssh can access the container of docker. If it cannot a workaround is needed such as adding docker img before each command, where img is the image name.

Additionally the install should still fail as the docker image is not loaded there yet, as the decision how to load it still needs to be made.