Fixed some logic errors in the tests.
Fixed -4 < 0 to expect True instead of False in the code.

Removed the tests that logically state 2 + j < 0 == False because < cannot be done with complex numbers.

Removed the testPositiveString method because Strings can be compared to numbers and the comparison is not supposed to be within the scope of the tests as only the value mattered.